Reorganize folders, make sure directory.json exists
