+++ index.js
@@ -4,9 +4,9 @@ var __importDefault = (this && this.__importDefault) || function (mod) {
 };
 Object.defineProperty(exports, "__esModule", { value: true });
 // check that all environment variables are set
-const EnvChecker = require("./envChecker");
+const EnvChecker = require("./js/envChecker");
 new EnvChecker();
-const competition_1 = __importDefault(require("./competition"));
+const competition_1 = __importDefault(require("./js/competition"));
 const competition = new competition_1.default();
 const discordClient = competition.makeDiscordClient();
 const webServer = competition.makeWebServer(discordClient);
