+++ js/competition.js
