+++ js/competition.ts
