+++ js/envChecker.js
