+++ js/envChecker.ts
@@ -9,6 +9,18 @@ module.exports = class EnvChecker {
             throw new Error("No .env file found!");
         }
 
+        // check that competitions folder exists
+        if (!fs.existsSync("competitions")) {
+            // make it
+            fs.mkdirSync("competitions");
+        }
+
+        // check that competitions/directory.json exists
+        if (!fs.existsSync("competitions/directory.json")) {
+            // make it
+            fs.writeFileSync("competitions/directory.json", "{}");
+        }
+
         // check that all required variables are set
         const requiredVariables = [
             "token",
