+++ js/scrambler.ts
